Ignore responses to previous auctions (#1239)
Consider the following scenario:

* Auction A sends requests to partners 1, 2, 3
* Partners 1 and 2 send back responses in time
* Auction timeout is triggered; callback invoked
* Auction B sends requests to partners 1, 2, 3
* Then partner 3’s response to **Auction A** comes back
* This response is determined to have timed out based on the delta
  between the request timestamp and response timestamp
* Bid manager invokes the callback with the `timedOut` flag set to true
* This causes **Auction B** to time out!

In general, when a response comes back, the bid manager should discard
it if it isn’t a response to a bid in the current auction. This patch
introduces a check in `bidManager.addBidResponse` which does an early
check to make sure the `requestId` of the bid matches the `requestId` of
the currently running auction (if any), and bails if it doesn’t.