Update renderAd to replace ${AUCTION_PRICE} in adUrl (#1795)
* Modify `adUrl` instead of `url` with bid price

The `bid.url` parameter is not used to render the iframe for an ad that
returns a url but is being treated as such when the auction price is
being inserted. This commit inserts the auction price into `bid.adUrl`
which is used to render the ad.

* Use the key for adUrl instead of using url

Instead of renaming the parameter `adUrl` to `url` for the rendering
function, use `adUrl` to be consistent with the key value returned in
the bid object.