Rubicon adapter: get referrer from bidderRequest.refererInfo.referer; (#3087)
* DecodeUrl if url is encoded

Appnexus new module is passing the encoded url and sometimes Publisher are passing encoded url in the referrer parameter. 
This fix is a protection in case a module is update and start passing encode url. Our AE is not decoding and from some test I did, monetisation could be highly impacted.

* Update rubiconBidAdapter.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter.js

* Update rubiconBidAdapter.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter.js

* Update rubiconBidAdapter_spec.js

* Update rubiconBidAdapter_spec.js