Add type information to DSPX bidResponse (#5021)
* Add type information to DSPX bidResponse

Depending on the ad type other ad placements can be blocked or not.
For example a Skin Ad may require that all skys are blocked so they
don't overlap. For a regular display banner this isn't necessary.

* Add type to unit test

* Revert "Add type information to DSPX bidResponse"

This reverts commit 9403ceac4f636c6c3f8576dbfb18ef6d1b93ff16.

* Add type information to DSPX bidResponse

Depending on the ad type other ad placements can be blocked or not.
For example a Skin Ad may require that all skys are blocked so they
don't overlap. For a regular display banner this isn't necessary.

* Revert "Add type to unit test"

This reverts commit c08ce996a2317adb968033920770e0af1a80efa6.

* Add type to unit test